requirements.txt
asgiref==3.2.10
authlib==0.14.1
celery==5.0.5
coverage==5.0.4
cssselect==1.1.0
disco-py==0.0.14rc1
discord.py==1.5.1
django-admin-sortable2==0.7.8
django-autocomplete-light==3.8.2
django-colorfield==0.4.1
django-environ==0.4.5
django-extensions==3.1.0
django-markdownx==3.0.1
django-object-actions==3.0.1
django-redis==4.12.1
django-webpack-loader==0.7.0
django==3.1.3
factory_boy==2.12.0
furl==2.1.0
ipython==7.5.0
irc==17.1
lxml==4.5.1
markdown==3.3.3
Pillow==7.1.1
PyHamcrest==2.0.2
pytest==4.6.11
pytest-django==3.10.0
git+git://github.com/michaelelin/pytest-relaxed@00cb382#egg=pytest-relaxed
pyyaml==5.1.2
requests==2.22.0
rollbar==0.15.1
